add latent mode module mixin
